PackageManager: allow build-time disabling of components
Allow components to be specified as disabled at build time
(applied on boot).

This allows stock OTA components to be marked as disabled in
CM builds.


SystemUpdateService: enable service but lock its receivers [1/2]

Added a check for ensure that disabled components are not
re-enabled at runtime

Added code for forcing enable of previously disabled components

Signed-off-by: Shekhawat2 <anandsingh215@yahoo.com>
Signed-off-by: phenomenall <venkysaihot@gmail.com>